pyproject.toml
[project]
name = "efficiency-fin-soccer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.0",
    "coloredlogs>=15.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.66.0",
]
